Enable physical model shadows (with animation) and change elevation to a double (#9756)
Fixes https://github.com/flutter/flutter/issues/9342

+++ dev/benchmarks/complex_layout/lib/main.dart
@@ -66,7 +66,7 @@ class TileScrollLayout extends StatelessWidget {
           return new Padding(
             padding:const EdgeInsets.all(5.0),
             child: new Material(
-              elevation: index % 5 + 1,
+              elevation: (index % 5 + 1).toDouble(),
               color: Colors.white,
               child: new IconBar(),
             ),
